I began with the asciidoc version generated by Tim Fanelli from the Word document and after his significant cleanup work. That was located here:  https://github.com/WASdev/standards.jsr352.batch-spec
Below are the changes made from that point..
Issue numbers are from that copy
---
Assorted formatting cleanup stuff
Issue WASdev#4 - Missing section header 10.5 skews section number afterwards
-> Fixed by putting a blank line before section header

Issue WASdev#3 - Section numbers only go down 3 levels
-> Fixed by specifying :sectnumlevels: 5 in the main doc.  Note that this is an undocumented option Google found for me in a forum post from 2015 (https://discuss.asciidoctor.org/Section-Numbering-Levels-td3879.html)

Section 1 - License - lettered bullet list items are all 'a'
-> Fixed by changing to a,b,c,d

Section 8.2.6 Partitioning - XML sample not formatted, "Syntax:" not on its own line
-> Added blank lines before "Syntax:" and before the xml

Section 8.4.1 Split Termination Processing Incomplete - "Incomplete" was not part of the title
-> Removed line break between "Processing" and "Incomplete"

Section 8.7.1 - Strikethrough sentence shouldn't be bolded
-> changed asterisks around strikethrough text to pound signs (* to #)
This was mentioned in issue WASdev#1 along with a reference to section 9.6 missing a space which was already fixed.

Sections 8.6.1, 8.6.2, and 8.6.4 examples aren't formatted
-> Either move the open '<' out of column one or add [source,xml] before XML delimited by four dashes '----'

Under Section 8.8.81 in subsections "jobProperties Substitution Operator" and "partitionPlan Substitution Operator"
-> Reformatted XML with better indention.
-> Also removed an extraneous </properties> from the jobProperties example that had been there forever (in original)

Section 8.8.2 - examples of property substitution
-> formatting was honoring line breaks so name and value were on separate lines, merged 'em together so they fit on one line making it more readable

Section 8.9.4 - paragraphs aren't numbered as an ordered list so references to rule WASdev#2 is meaningless
-> Added ". " before paragraphs to trigger numbering and a ".. " before the note under item 2 to sublist as 'a'.

Section 8.9.5 - the word 'Undefined' should have been part of the title but was on the next line
-> removed line break
-> Also cleaned up formatting of JSL

Section 9.1.2 - At the end of the Batchlet interface there was an embedded image with text about batchlet stop processing
-> The image had an unresolved reference in it and just looked silly.  Made it into a TIP: with a link/ref to the Stop section.

Section 9.3.1 (@BatchProperty) - A 'Note' paragraph was in a gray box for no reason
-> Removed the space before 'Note' that made it an example

Section 9.3.1 (@BatchProperty) - The MyItemReaderImpl sample code block has an extraneous '\' after the class name
-> Removed it.  Wasn't in the original..

Section 9.3.2 - 'Injection' should be part of the title but is on the next line
-> Removed extraneous line break
-> Also cleaned up XML sample indention

Section 9.4.1.1 has an item list of JobContext and StepContext.  StepContext had a '+' at the end
-> put there to force a line break, there was no space between the 't' and the '+'.  Added one.

Section 9.5 - Partitioned/Concurrent sections need to be numbered with inner list ordered with letters
-> Changed number to the '. ' and '.. ' syntax with a '+' on blank lines to continue list paragraphs.

Section 10.9.9 (JobExecution) - getCreateTime has an extraneous 'updated' in the javadoc
-> Removed it (this was in the original spec too), mentioned in Issue WASdev#1

Lifecycle flows have some errors
-> In Partitioned Batchlet Processsing, in the original PDF steps 8, 10, and 12 are indented but numbered
as part of the main sequence (showing they are part of the conditional in the line above).  In the asciidoc
in lines 8 and 10 we indented these with a '..' which threw off the numbering, so I removed the extra dot.
This left things numbered right, but we lost the indentation for the conditional (this step is the 'then' of an 'if').
I used five non-blank-space elements ({nbsp}{nbsp}{nbsp}{nbsp}{nbsp}) to scoot it over a little.
-> In Regular Chunk Processing step 9.b.i was missing a space after the dots so it got folded into 9.b.  Added a space
-> Note also that four layers in (9.b.i.1 in the original) uses letters so 9.b.i.A.  Didn't try to fix this.
-> In Partitioned Chunk Processing there's a comment after step 5 but the '+' to put it on a separate line was missing
a leading space so it was taken literally.  Added a space.
-> Also items 5.e.v and vi were run into the prior item due to missing spaces after the dots, added a space
-> And some more non-blank-spaces added to get indention
-> In "Chunk with Listeners (except RetryListener)" a whole block was indented one too many levels..backed it out one
-> In Chunk with RetryListener one line split throwing the count off, merged that..and then a lot of the nbsp stuff
to try to get the indentations right
-> In Chunk with Custom Checkpoint Processing more indentation level problems

Throughout - numerous reuses of the app-listing block identifier
-> Tacked on the name of the block it was identifying (e.g. app-listing.ItemReader.java)
-> Also fixed the PartitionAnalyzer block name which mis-identified it as a PartitionCollector

Throughout - XML sample indention needs tidying up
-> Some are specified above, but touched up a bunch of other ones

Throughout - XML blocks aren't tagged that way..seems to default properly
-> but thought I'd be complete and mark 'em all...

Sign off

Signed-off-by: David Follis <follis@us.ibm.com>